+++ .changeset/polite-adults-sip.md
@@ -2,4 +2,4 @@
 "@sit-onyx/playwright-utils": minor
 ---
 
-feat: implemented emit spy helper functions: `createEventSpy` and `expectEventCall`
+feat: implemented emit spy helper functions: `createEmitSpy` and `expectEmit`
