+++ packages/playwright-utils/src/index.ts
@@ -1,4 +1,4 @@
-export * from "./events/index.js";
-export * from "./events/types.js";
+export * from "./emits/index.js";
+export * from "./emits/types.js";
 export * from "./screenshots/index.js";
 export * from "./screenshots/types.js";
